Rewind: App Page Continued
Made some adjustments to site.js. Decided to use an array instead of a string (eventually, I joined the array back into a string for the final result) to better fit the challenge description.

Misc. adjustment: I set up the project to use the Chrome debugger to help step through the code as it is executing.